Change `unknown` field type in `m_Header` struct from `long` to `uint32_t` for cross-platform compatibility

+++ include/esptk/tes3record.h
@@ -23,7 +23,7 @@ private:
   {
     char type[4];
     float version;
-    long unknown;
+    uint32_t unknown;
   } m_Header;
 };
